Add MD5 hash algorithm
This implementation is pretty cleanroom based on the spec.  I focused
more on code size than on speed and even with needing to byteswap it's
still noticeably faster than SHA1.  It should be noted that I knew
nothing about hash algorithms before starting this, I just didn't like
the existing implementations of MD5 in JS and none of them fit into SJCL
neatly.