toBits(): Round up len to nearest byte instead of radix.
When toBits() is called without specifying a length, calculate the
length with bitLength(), which rounds up to the nearest byte instead of
multiplying limbs by radix, which rounds up to the nearest radix.  This
fixes an issue where converting a 256-bit value to a bitArray and then
calling bitArray.bitLength() on it, gave 264.